requirements.txt
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
reportlab==4.2.2